# docs/swagger.yaml
basePath: /api/v1
definitions:
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.AmbienteResponseDTO:
    properties:
      altura:
        type: number
      comprimento:
        type: number
      descricao:
        type: string
      id:
        type: integer
      largura:
        type: number
      nome:
        type: string
      orientacao:
        type: string
      tempo_exposicao:
        type: integer
      tipo:
        type: string
    type: object
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.CreateAmbienteDTO:
    properties:
      altura:
        description: em centímetros
        type: number
      comprimento:
        description: em centímetros
        type: number
      descricao:
        type: string
      largura:
        description: em centímetros
        type: number
      nome:
        type: string
      tempo_exposicao:
        description: em horas
        type: integer
      tipo:
        description: 'Ex: "interno", "externo", "húmido", "seco"'
        enum:
        - interno
        - externo
        - húmido
        - seco
        type: string
    required:
    - altura
    - comprimento
    - largura
    - nome
    - tempo_exposicao
    - tipo
    type: object
...
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.PaginatedResponse:
    properties:
      data:
        items:
          type: integer
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.UpdateAmbienteDTO:
    properties:
      altura:
        type: number
      comprimento:
        type: number
      descricao:
        type: string
      largura:
        type: number
      nome:
        type: string
      tempo_exposicao:
        type: integer
      tipo:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API for the Cultivo app.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Cultivo API
  version: "1.0"
paths:
  /api/v1/ambientes:
    get:
      description: Retorna uma lista paginada de todos os ambientes cadastrados
      parameters:
      - description: 'Número da página (padrão: 1)'
        in: query
        name: page
        type: integer
      - description: 'Limite de itens por página (padrão: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.AmbienteResponseDTO'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Lista todos os ambientes com paginação
      tags:
      - ambiente
    post:
      consumes:
      - application/json
      description: Cria um novo ambiente com os dados fornecidos
      parameters:
      - description: Dados do Ambiente
        in: body
        name: ambiente
        required: true
        schema:
          $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.CreateAmbienteDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.AmbienteResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cria um novo ambiente
      tags:
      - ambiente
  /api/v1/ambientes/{id}:
    delete:
      consumes:
      - application/json
      description: Deleta um ambiente existente
      parameters:
      - description: ID do Ambiente
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Deleta um ambiente
      tags:
      - ambiente
    get:
      consumes:
      - application/json
      description: Retorna os detalhes de um ambiente específico
      parameters:
      - description: ID do Ambiente
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.AmbienteResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Busca um ambiente por ID
      tags:
      - ambiente
    put:
      consumes:
      - application/json
      description: Atualiza um ambiente existente com os dados fornecidos
      parameters:
      - description: ID do Ambiente
        in: path
        name: id
        required: true
        type: integer
      - description: Dados do Ambiente para atualização
        in: body
        name: ambiente
        required: true
        schema:
          $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.UpdateAmbienteDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gitea_paulojamil_dev_br_paulojamil_dev_br_cultivo-api-go_internal_domain_dto.AmbienteResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualiza um ambiente
      tags:
      - ambiente
...
swagger: "2.0"